Day 8 completed
This one was challenging in terms of index management. I always fall into the trap of "wait is the first one x, or y. Which one is collumns?" but keeping a consistent logic of "first means row" and sticking to it works. In this case, it doesn't REALLY matter if I am doing row or column in a given check, aslong as it is consistent, so I just convince myself of one of them and move on